requirements.txt
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
#blinker==1.4
cachetools
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
#click==7.1.1
cryptography==2.8
cytoolz==0.10.1
dapr==1.2.0
dapr-client==0.4.0b1
dapr.ext.grpc==1.2.0
datetime
dropbox==9.5.0
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
Flask==2.3.3
#Flask==1.1.1
Flask-Dance==3.0.0
Flask-Dapr==1.2.0
Flask-Mail==0.9.1
google-auth
google-auth-oauthlib
google-auth-httplib2
gunicorn==21.2.0
hexbytes==0.2.0
idna==2.9
#importlib-metadata==1.5.2
ipfshttpclient==0.4.12
#itsdangerous==1.1.0
#Jinja2==2.11.1
jsonify==0.5
jsonschema==3.2.0
kafka-python
lru-dict==1.1.6
#MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.7.19
oauthlib==3.1.0
parsimonious==0.8.1
protobuf==3.13.0
pycparser==2.20
pycryptodomex
pycryptodome==3.9.7
PyJWT==1.7.1
# pypiwin32==223
pyrsistent==0.16.0
python-dotenv==1.0.0
# pywin32==227
requests==2.23.0
requests-oauthlib==1.3.0
rlp==1.2.0
sendgrid
six==1.14.0
toolz==0.10.0
typing-extensions==3.7.4.1
urllib3==1.25.8
URLObject==2.4.3
varint==1.0.2
web3==5.7.0
websockets==8.1
#Werkzeug==1.0.0
zipp==3.1.0